requirements.txt
flask
opencv-python
pip-chill
python-dotenv
